+++ feature/timer/src/main/java/org/sopt/timer/settimer/clipselect/ClipSelectAdapter.kt
@@ -1,4 +1,4 @@
-package org.sopt.timer.settimer
+package org.sopt.timer.settimer.clipselect
 
 import android.content.Context
 import android.view.LayoutInflater
