+++ feature/timer/src/main/java/org/sopt/timer/settimer/timepicker/PickerViewHolder.kt
@@ -1,5 +1,6 @@
-package org.sopt.timer.settimer
+package org.sopt.timer.settimer.timepicker
 
+import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import org.sopt.mainfeature.R
 import org.sopt.timer.databinding.ItemNumberPickerBinding
